convert uint, float64, to int

+++ syncstorage/BSO.go
@@ -2,9 +2,9 @@ package syncstorage
 
 // ref: https://docs.services.mozilla.com/storage/apis-1.5.html#basic-storage-object
 type BSO struct {
-	Id        string  `json:"id"`
-	Modified  float64 `json:"modified"`
-	Payload   string  `json:"payload"`
-	SortIndex uint    `json:"sortindex"`
-	TTL       uint    `json:",omitempty"`
+	Id        string `json:"id"`
+	Modified  int    `json:"modified"`
+	Payload   string `json:"payload"`
+	SortIndex int    `json:"sortindex"`
+	TTL       int    `json:",omitempty"`
 }
